feat: move json files into data folder
