+++ src/sections/Innovation.tsx
@@ -1,5 +1,5 @@
 import React, { useEffect, useState } from "react";
-import data from "../data.json";
+import data from "../data/data.json";
 import { useGlobalContext } from "../GlobalContext";
 import { StickyItem } from "../types/global";
 import Card from "../components/Card";
